# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": [
    "final-cse-341-group.onrender.com"
  ],
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "name": "apiKey",
      "in": "header"
    }
  },
  "paths": {
...
    "/authors/": {
      "post": {
        "tags": [
          "Authors"
        ],
        "summary": "Create a new author",
        "description": "Endpoint to create an author",
        "parameters": [
          {
            "name": "Author",
            "in": "body",
            "description": "Author Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Mario Vargas Llosa"
                        }
                      }
                    },
                    "birthdate": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "date"
                        },
                        "example": {
                          "type": "string",
                          "example": "1936-03-28"
                        }
                      }
                    },
                    "nationality": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Peruvian"
                        }
                      }
                    },
                    "biography": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Lorem ipsum dolor sit amet"
                        }
                      }
                    },
                    "website": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "url"
                        },
                        "example": {
                          "type": "string",
                          "example": "https://mariovargasllosa.net"
                        }
                      }
                    },
                    "socialMediaLinks": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "Twitter"
                                    }
                                  }
                                },
                                "url": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "format": {
                                      "type": "string",
                                      "example": "url"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "https://www.twitter.com/mariovargasllosa"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        },
                        "example": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "Twitter"
                              },
                              "url": {
                                "type": "string",
                                "example": "https://www.twitter.com/mariovargasllosa"
                              }
                            }
                          }
                        }
                      }
                    },
                    "contactInformation": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "Email"
                                    }
                                  }
                                },
                                "value": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "mariovargasllosa@example.com"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        },
                        "example": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "Email"
                              },
                              "value": {
                                "type": "string",
                                "example": "mariovargasllosa@example.com"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get all authors",
        "description": "Endpoint to get all authors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get an author by ID",
        "description": "Endpoint to get an author by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Authors"
        ],
        "summary": "Update an author by ID",
        "description": "Updates an existing author in the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the author to update"
          },
          {
            "name": "Author",
            "in": "body",
            "description": "Author Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Mario Vargas Llosa"
                        }
                      }
                    },
                    "birthdate": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "date"
                        },
                        "example": {
                          "type": "string",
                          "example": "1936-03-28"
                        }
                      }
                    },
                    "nationality": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Peruvian"
                        }
                      }
                    },
                    "biography": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Lorem ipsum dolor sit amet"
                        }
                      }
                    },
                    "website": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "url"
                        },
                        "example": {
                          "type": "string",
                          "example": "https://mariovargasllosa.net"
                        }
                      }
                    },
                    "socialMediaLinks": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "Twitter"
                                    }
                                  }
                                },
                                "url": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "format": {
                                      "type": "string",
                                      "example": "url"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "https://www.twitter.com/mariovargasllosa"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        },
                        "example": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "Twitter"
                              },
                              "url": {
                                "type": "string",
                                "example": "https://www.twitter.com/mariovargasllosa"
                              }
                            }
                          }
                        }
                      }
                    },
                    "contactInformation": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "Email"
                                    }
                                  }
                                },
                                "value": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "mariovargasllosa@example.com"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        },
                        "example": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "Email"
                              },
                              "value": {
                                "type": "string",
                                "example": "mariovargasllosa@example.com"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Authors"
        ],
        "summary": "Delete an author by ID",
        "description": "Endpoint to delete an author by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}